Split up class Validator, separating concerns
