+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/data/IsoDocumentParsed.kt
@@ -6,7 +6,7 @@ import at.asitplus.iso.IssuerSignedItem
 import at.asitplus.wallet.lib.iso.MobileSecurityObject
 
 /**
- * Intermediate class used by [at.asitplus.wallet.lib.agent.Validator.verifyDocument] when parsing an ISO document,
+ * Intermediate class used by [at.asitplus.wallet.lib.agent.ValidatorMdoc.verifyDocument] when parsing an ISO document,
  * and also in [at.asitplus.wallet.lib.agent.VerifierAgent.verifyPresentationIsoMdoc].
  */
 data class IsoDocumentParsed(
