+++ vck/src/commonTest/kotlin/at/asitplus/wallet/lib/agent/AgentRevocationTest.kt
@@ -136,7 +136,7 @@ class AgentRevocationTest : FreeSpec({
         }
         result.shouldBeInstanceOf<Issuer.IssuedCredential.VcJwt>()
 
-        val vcJws = Validator().verifyVcJws(result.signedVcJws, verifierKeyMaterial.publicKey)
+        val vcJws = ValidatorVcJws().verifyVcJws(result.signedVcJws, verifierKeyMaterial.publicKey)
         vcJws.shouldBeInstanceOf<Verifier.VerifyCredentialResult.SuccessJwt>()
         val credentialStatus = vcJws.jws.vc.credentialStatus
         credentialStatus.shouldNotBeNull()
